Keymap Revamp with Accents and and Hand-Pairing
Accents: Added Umlaut (¨), Grav (`), Egu (´), and Circumflex (ˆ)

Hand-Pairing: Made it so that symbols are typed by hand not holding layer key. Did this by switching Left and Right layer and moving a lot of the symbols around. 

Also: Reduced emphasis on upper row as I find it more comfortable to reach the lower one than the upper one